painter changes
