+++ src/styles.css
@@ -1,6 +1,23 @@
 body{
 	margin: 0;
+	overflow: hidden;
 }
 canvas{
 	box-shadow: 0 0 5px gray;
+	/*width: 300px;*/
+	/*height: 300px;*/
+	/*float: left;*/
+}
+
+.main{
+	/*display: flex;*/
+}
+.toolset{
+	float: right;
+	display: flex;
+	flex-direction: column;
+	padding: 5px;
+}
+.sketch-picker{
+	width: 115px !Important;
 }
