+++ tsconfig.json
@@ -17,6 +17,7 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
+    "noImplicitThis": false,
     "jsx": "react"
   },
   "include": [
